Fix some lints

+++ isla-lib/src/bitvector/b129.rs
@@ -392,8 +392,8 @@ impl BV for B129 {
 
     fn set_slice_int(int: i128, n: u32, update: Self) -> i128 {
         assert!(update.len <= 128);
-        let mask = !bzhi_u128(u128::max_value() << n, n as u32 + update.len());
-        let update = (update.bits as u128) << n;
+        let mask = !bzhi_u128(u128::max_value() << n, n + update.len());
+        let update = update.bits << n;
         ((int as u128 & mask) | update) as i128
     }
